# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bondhon Matrimony API",
        "description": "Full-stack matrimony platform API. All protected endpoints require Bearer token from Sanctum.",
        "contact": {
            "email": "admin@bondhon.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "7ca6fc7989202054d29707fef91a2d4a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "gender",
                                    "profile_created_by"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Rahim Uddin"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "rahim@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "male",
                                        "enum": [
                                            "male",
                                            "female"
                                        ]
                                    },
                                    "profile_created_by": {
                                        "type": "string",
                                        "example": "self",
                                        "enum": [
                                            "self",
                                            "parents",
                                            "siblings"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "operationId": "9d0bb5c770cbec19b0b9566606ef4adb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "rahim@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "403": {
                        "description": "Account banned or inactive"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user (revoke token)",
                "operationId": "3a27a9e02ceb521eb3f98c8d2394aef9",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get authenticated user details",
                "operationId": "7edb820901b8b3a880fc4d7a0a7290fb",
                "responses": {
                    "200": {
                        "description": "User details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Email Verification"
                ],
                "summary": "Verify user email address",
                "operationId": "d3488e1ddfc474c63037c03f7700c4e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email verified successfully"
                    },
                    "403": {
                        "description": "Invalid or expired verification link"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/auth/email/resend": {
            "post": {
                "tags": [
                    "Email Verification"
                ],
                "summary": "Resend email verification link",
                "operationId": "8868dc08613cf7995f711a30c390a484",
                "responses": {
                    "200": {
                        "description": "Verification link sent"
                    },
                    "400": {
                        "description": "Email already verified"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token. Obtain via POST /api/v1/auth/login",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}